Fix compatibility with PHP 5.3 and 5.4
In contrast to what is "advertised", the test server code as was had a minimum PHP requirement of PHP 5.5.
This prevented the tests from running/passing on PHP 5.3/5.4 in combination with GH actions.

This commit fixes that so we can release one last version of Requests with _proven_ support for PHP 5.3-5.4.

Fixes made:
* Replace short arrays with long arrays. (introduced in PHP 5.4)
* Replace use of `yield`. (introduced in PHP 5.5)
* Polyfill the `http_response_code()` function (introduced in PHP 5.4).
* Take into account that `JSON_PRETTY_PRINT` may not be available (introduced in PHP 5.4).
* Give `html*()` functions an encoding to use as the default value changed in PHP 5.4.